Moving data layers to layer control
